Fix slickGoTo not using animations
Added a new option to `changeSlide` to handle a specific index.  InnerSlider now
uses that new functionality and also skips the `this.update` call in
`componentWillReceiveProps` which preserves the animation.

The use case I'm trying to solve for is having paging buttons that live outside
the Slider component (design requirement for application I'm working on).  This
seemed like the least invasive way to do it.